убрать из HTTPTS nandler and adapter в свои пакеты
